Upate mattresses
Display an unallocated mattress in the mattress list with an amount of the account total net balance minus the sum of the other mattress amounts.

Properly format the dollar amounts in the mattress list and transfer form.

Update the send function to allow for mattress to be requested sequentially so they stay in alphabetical order.

Closes #22